Retry flaky tests
Summary:
Adds an option to the test runner that specifies how many times we
should retry flaky tests.  This defaults to 2.

We don't really know whether a test is flaky or not ahead of time,
so this really just specifies how many times we'll retry a failing
test.

We'll only perform this retry behavior if the test instance has
a method that allows setting the current attempt number.  We do
this because most of our tests derive an output directory from
the current test configuration; we need to augment that path with
the attempt number in order for the subsequent attempt to be able
to run in a clean environment.

I've enabled this for python and php tests.

A failing test will still print out its failure reason on each attempt,
along with the current attempt number, but if it passes on a subsequent
attempt, the overall test run status will reflect only its passing
result.

We could potentialy track a number of flaky tests and report it at the end,
but I just want to get this in so that our OSS CI needs less manual poking.

In addition to the above, tweak how we collect and report failing tests
in travis and circle, so that we don't overwhelm their respective UIs with
log data.

I've also temporarily disabled the watchman java tests; we'll follow up
and fix those, but I just want to get the CI into a default green state.

Closes https://github.com/facebook/watchman/pull/473

Differential Revision: D5101754

Pulled By: wez

fbshipit-source-id: c2ee1485611e3566e4ec5b74e97273a4612038af