rename watchman_stat to FileInformation
Summary:
This is part of a series of diffs to improve our portability with
windows.  They also have the nice side effect of addressing some performance
and reliability issues that we're tracking on github.

An important part of this diff stack is removing our nasty stat() hacks; this
diff creates a FileInformation type for this purpose.

FileInformation holds a subset of the struct stat information as well as some
additional win32 specific information.

It introduces a handful of helper methods for testing whether the
FileInformation references a symlink, dir or regular file.  This is important
because the sys/stat.h that is provided by MSVCRT does not have any way to
indicate that the file is a symlink.

Reviewed By: simpkins

Differential Revision: D5091465

fbshipit-source-id: cbdbd9edc87fba005e79f492001be93abd344ac9