use generic build info from fb build system for version
Summary:
We always shows 4.9.0 for our builds, but the version number doesn't mean very much to us with our more frequent deploys than public releases.

This diff allows us to pull the build info from the fbcode build system and use that instead.

The version number shows up as the build time instead of 4.9.0, and the buildinfo property is set to the revision hash.

Reviewed By: simpkins

Differential Revision: D5107953

fbshipit-source-id: 0bf19f76d5bbc351e324adba6cc3451baf537532