add moral equivalent to strict lstat
Summary:
adds a getFileInformation(const char *path) function that sets a
strict name checking flag.  This asserts that the path being examined is the
canonical name on the filesystem before allowing the stat info to be returned.

This also removes setting the st_dev field to anything useful on Windows.
The best we could do is compute this from the drive letter, but we don't
always know the drive letter when getting file information (fstat equivalent)
without performing an expensive additional check to get the associated
path.

Reviewed By: farnz

Differential Revision: D5091546

fbshipit-source-id: e0a39e427abb1993a19a305ec3c72d1c0214030c