improve handling of very large queries
Summary:
The -j option would only read at most two chunks from
the input file.  The default chunk size is 1MB, so this meant
that we'd only successfully be able to read in commands smaller
than 2MB in size.

This diff changes the if to a while so that we keep trying
until we exhaust the input stream.

In addition, when computing the bser read buffer size on the server side, we'd
under-estimate the required size and then never be able to satisfy reading all
of the data, erroring out just before the end of the buffer.

On windows, running this test found a heap corruption bug; at some point
in the past, we moved away from holding a separate data buffer in our
windows stream class, but we didn't clean up all the free() calls.
There was one remaining here that could crash the process if we destroyed
a stream before we had flushed all pending write data.

I've had to disable this new test on windows because we have a blocking
problem with large command buffers.  I tried to resolve this today,
but its a thorny problem and best taken care of in its own diff.
The problem is that we need to flush out the partially written command
buffer before we can sanely block on the read.

Differential Revision: D5128607

fbshipit-source-id: 7f682d4cf2403aec76951245ece0445bb7e40191