remove calls to w_lstat
Summary:
These have been replaced by calls to getFileInformation.

I took out the stat() call from the pending test, because it wasn't
really needed and because the actual contents of the __FILE__ string
is at the mercy of the user/compiler, and I was too lazy to write
code to make things handle this correctly in a test that didn't
really need it.

While testing this, I found that we also need to add in the strict
name checks for getFileInformation on the mac.

Reviewed By: simpkins

Differential Revision: D5094991

fbshipit-source-id: 96e7a7469f09c495a743cd74f930fdff49cee3cb