add realPath, successor to w_realpath
Summary:
This uses the getOpenedPath() functions added earlier
in the stack, and thus has cleaner and more portable error
handling semantics.

I'll follow up with diffs to remove the remaining w_realpath
calls.

Note that this diff allows using "." to reference the cwd
on windows, which is something that wasn't working previously.

Reviewed By: farnz

Differential Revision: D5100956

fbshipit-source-id: 093af4f9ab8bd85be9cda888e0cd93d5798d41f9