cut opendir code over to os independent api
Summary:
This allows us to remove some of the older strict file opening
code, and makes the code look a little bit more uniform overall.

I've made the equivalent change to the win32 version of this, and
found that I should always enable BACKUP_SEMANTICS for the win32
handle options to allow operating on a directory handle; it's safe
to always add that flag even for files.

Reviewed By: farnz

Differential Revision: D5100961

fbshipit-source-id: 504f775b7a00dd890142a5c2dda3c726d0174c5a