enable watch_symlinks on win32
Summary:
Remove the preprocessor directive that restricted this to !win32
and fixup the code to use the new readSymbolicLink(path) function.

Fixup off by one in baseName, dirName.  Both of these functions would start
reading at the end pointer rather than just before it.  This typically worked
out OK, but if that end position contained a slash character, we might try to
construct a string with a bogus length.  This is deteted at runtime via the
checked_len function which throws an exception and cancels the watch.

Reviewed By: simpkins

Differential Revision: D5101087

fbshipit-source-id: eb77f731539328ced4ddd844715a5891304a30a3