merge Win32Handle into FileDescriptor
Summary:
I've kept Win32Handle.cpp and some of the functions it
holds for the moment; since I have a number of modifications in
a stack of changes that is pending review, I'm anticipating that
resolving merge conflicts across a rename will be a PITA, so I'm
deferring moving those around until after the stack is landed.

This is a pretty straight forward change with just a little bit
of fan-out to the heavier weight stream layer.

Reviewed By: simpkins

Differential Revision: D5124899

fbshipit-source-id: 14b9cd76ec67881c5d26edc5ee98f31e75c453d4