gcc 4.8 on travis miscompares std::error_condition
Summary:
This appeared to be an issue with the implicit conversion of
std::error_condition-enabled enum values up to std::error_code values with the
version of gcc 4.8 that was in use on the travis hosts.  Adding an explicit
call to convert them resolves the issue.

The manifestation of the issue is that deleted files were not recognized as
deleted and the resultant view of the filesystem was rendered incorrect.

Closes https://github.com/facebook/watchman/pull/476

Differential Revision: D5180078

Pulled By: wez

fbshipit-source-id: 87bc04848f16a95310437b9d0cd47cb77c9acc54