handle readdir errors in the iothread
Summary:
These don't seem to be possible, or at are least vanishingly rare,
on posix systems, but crop up in the git test suite on windows
(https://github.com/facebook/watchman/issues/461)

Uncaught exceptions in the IO thread result in the watch being
cancelled, so we need to do better.

This diff catches exceptions raised by readDir and adds the path
back to the pending list.   The theory is that these exceptions
can only happen if the directory has been removed while we were
reading it, so the subsequent stat should get an ENOENT equivalent.

The danger here is that we get stuck in a loop trying to stat these
forever, so we may need to do some immediate processing here instead.

I've also added ERROR_DIRECTORY to the error category stuff so that
we recognize it as being equivalent to ENOTDIR.

Closes https://github.com/facebook/watchman/pull/477

Reviewed By: simpkins

Differential Revision: D5183409

Pulled By: wez

fbshipit-source-id: ca2ee4340f617c13fd06c198b6d412c34091bf67