watchman: nullptr check for stringpiece construction
Summary:
When using source control aware queries, it is possible that
the child process won't emit any output.  When that happens, our ChildProcess
code can return a `w_string(nullptr)`.  When this is passed to the `w_string_piece`
constructor, we would blindly deref the nullptr by calling the `.data()` and `.size()`
methods of `w_string`.

This diff:

1. Adds an unambiguous overload for `w_string_piece::w_string_piece` to construct from a `w_string` that may hold a nullptr value.
2. Adds a couple of unit tests for this case
3. Makes `w_string::data()` and `w_string::size()` throw when used on an instance that has a `nullptr` value.
4. Removes a redundant copypasta line from Mercurial.cpp

Reviewed By: simpkins

Differential Revision: D5178006

fbshipit-source-id: bf8e55c88e3d2a41c3ce134a0583d2f49b004ef0