Improve tests for dirName, baseName and fix corner case
Summary:
It's technically unspecified whether `e_ - 1 < s_` is true if `e_ == s_` and s_
points to the beginning of the memory block allocated for the string.

Further "[watchman] enable watch_symlinks on win32" did not add tests for the
fixes it made to dirName and baseName - add test cases to match.

Reviewed By: simpkins

Differential Revision: D5174324

fbshipit-source-id: cf91e90d3100187bc6bb5b6b34b9e360b671870b