avoid deadlock in trigger-del
Summary:
Some refactoring in the windows watcher highlighted this
potential problem; we hold a lock on the triggers and then try to
join a thread.  Depending on the state of the thread, this can deadlock.

The resolution is to move the join outside of the critical section
that erases the trigger from the locked map.

Reviewed By: simpkins

Differential Revision: D5209182

fbshipit-source-id: 92a73ec32db761d3e47430c030cd8e9411683d04