add w_string::empty and use it estimateSizeNeeded/toAppend
Summary:
I was debugging a problem where a w_string had been moved away,
leaving it as a nullptr w_string.  When logging or building strings with
this, our recent ensureNotNull check was triggering an assertion and
we'd fall over (with a garbage stack trace as fixed in the diff prior to this).

To allow safe logging of nullptr strings I've added an .empty() method that will
return true (and that behaves as expected for non-empty strings) and use this
as a portable way to tell if a generic stringy object has no content.

Reviewed By: simpkins

Differential Revision: D5209180

fbshipit-source-id: 91782ed131717f27aebb3a86edf153c2bdccd51d