improve IO scheduling around deletes on win32
Summary:
we've had a couple of reports that watchman interferes with
recursive directory removes on windows.

Looking at the code, we were pinging the IO thread for pretty much
each packet of changes coming from the OS.  The theory is that watchman
would get notified that the parent directory had a delete while the
remover was working through the contents.  Watchman would then try to
open handle(s) in that portion of the tree, accidentally blocking
the remove.

This diff seeks to improve the situation by:

1. Batching the group of notifications together more frequently.
   This is done by changing the timeout on the ReadDirectoryChangesW
   call.  When we have nothing in the batch we use a non-zero timeout.
   When we have something in the batch we use a 30ms timeout.
   The batch is published to the IO observer thread whenever we timeout the
   ReadDirectoryChangesW call.  This should reduce the chances that we'll
   race with a busy recursive mutation operation.
2. Take advantage of child path pruning in the pending list.  Looking
  at the flags in the notification struct from the OS, we can tell
  whether the path is going away (deleted or renamed away).  In those
  case we can safely unilaterally set the RECURSIVE flag on the path
  (normally we wouldn't want to, as a simple change on a dir may not
  warrant a full recursive examination under that path).   The recursive
  flag allows the pending list to collapse redundant paths out of its
  IO plan.  eg: if "foo/bar" and "foo" are both added with the recursive
  flag, the resultant pending list is just "foo" recursive.  For a
  recursive tree remove we can therefore transform a long list of
  discrete getFileInformation calls into a single operation on the
  root of that tree.   In addition, we can now benefit from the
  cached file information made available at readDir time.

Note that using the cached file info from readDir results in the nlink
field being set to 0 in some cases on Windows, and requires an adjustment
in the integration tests.  I'm mostly leaning towards this being "OK",
because it results in better overall perf, but it is technically a
regression and non-deterministic behavior.  It may be better to just
remove nlink as an available field on windows.  This field is difficult
to use in practice anyway.  I need to double check our behavior around
this field on linux and mac where were have similar optimizations.

Reviewed By: simpkins

Differential Revision: D5209181

fbshipit-source-id: 2aea3986f458235fa96181818350a4c9da693b1a