recrawls no longer recreate the view instance
Summary:
The primary problem I'm shooting to address here is what appears to be a race
condition with queries being placed at the same time as a recrawl is detected.
When our query lock was giant and always exclusive this was never a problem,
but in moving to finer grained locking we've opened ourselves up to this.

The race is around the view instance being recycled.  This occurred in some
thorny code that originated from some clever C code when it was originally
added and that has morphed into some tricksy placement new/placement delete
code in the codebase today.  Over time the members touched by the
tearDown/reinit flow have been reduced in volume, and we've reached the point
where we no longer require or benefit from this complication.

This diff removes the view reset at recrawl, which should eliminate all
possible recrawl/teardown related races.

It has the following implications:

* Recrawls no longer directly lead to a fresh instance result set.
  This is a positive move for all consumers; it changes the
  recrawl behavior from causing a higher latency and very large
  response to a higher latency response with smaller volume.
  This extends to triggers: we'll only run the trigger if
  we've detected files as having changed, whereas we would
  previously always run the trigger again on recrawl.

* Sync cookies were traditionally completely abandoned by a
  recrawl.  We can now abort them at recrawl time (actually,
  at the end of the recrawl) so that waiters are signalled
  about the issue rather than having to fully timeout.

  syncToNow will catch the cookie abort and retry with
  another cookie until the total timeout period has elapsed.

The heart of this change is to move the `handleShouldRecrawl` method from being
acted on in the kernel notification thread to the IO/observation thread in the
InMemoryView class.  When scheduling a recrawl, the recrawlInfo is updated to
reflect the request, and then the IO thread is pinged (via the new
`wakeThreads` method) to wake up and perform the recursive crawl.

I'm going to make a follow up diff to move the members from the `inner` struct
up to the main level in the `root`; I didn't want to inflate this diff with
that rather mechanical change.

Reviewed By: simpkins

Differential Revision: D5132237

fbshipit-source-id: 6b1dcb1cbb90beb690a0f13ea0c6bf0c89be0695