fixup subscribe test on win32
Summary:
The recent change to the watcher latency has exposed some
timing sensitivity in this test.   Let's remain true to the spirit
of the test by setting the config option for the latency to 0.

This is "OK" because regardless of the setting of the latency parameter,
the results will always be eventually consistent.

Reviewed By: simpkins

Differential Revision: D5365893

fbshipit-source-id: 3d627725a45463186be82bb34483e969aa0ca131