Enable Python client to be used as a context manager
Summary:
Currently, the Python client relies on __del__ to clean up
the active transport. Although, callers can explicitly call
close() if they want.

__del__ in Python is fundamentally dangerous for many reasons
and should not be used lightly. For example, exceptions during
__del__ are ignored and circular references on objects with
__del__ may result in deferred garbage collection. In some
scenarios, this can lead to resource "leaks."

Python's context managers are the preferred mechanism to
manage lifetimes of primitives. So this commit teaches the
"client" class to work as a context manager. When the context
manager is entered, the transport is activated. When it exits,
the transport is destroyed.
Closes https://github.com/facebook/watchman/pull/505

Differential Revision: D5496531

Pulled By: sid0

fbshipit-source-id: e0c5952b3e050901fbbd3bc3d50d0ed572fefec8