Fix failing async test case.
Summary:
Fixed an async test case where the test times out since the clock has not
changed (and thus, no subscription events are generated). The test is to make
sure that subscriptions are not lost after the 'debug-recrawl' operation. We
simply create a new file so that the clock increments. The intent of the test
is still achieved.

Reviewed By: wez

Differential Revision: D5726734

fbshipit-source-id: 693418c0dc3f32b6d679b9a3c4ebfcf8e8a24257