Implement ucred credentials on solaris/illumos in UnixStream class.
Implement ured credantials using solaris/illumos ucred functions.
SO_RECVUCRED is used as ifdef since thats the define used by
getsockopt on solaris to recive creds from sockets so it
follows the pattern the rest of the code uses, it is
not used in the code since the size of the cred structure
can change between boots (see man ucred_get) so malloc has to
be used and getpeercred fixes all this for us.
.