Fix problem where watch implicitly created.
Summary:
A user that runs the following commands will implicitly create a watch:
- state-enter
- state-leave
- subscribe
- trigger
- watch
This change will limit the watch creation to the watch command (explicit).
This will make behaviors more intuitive and prevent future issues.
We separate commands for the resolve_root and resolve_root_create - as a defensive measure.

Reviewed By: wez

Differential Revision: D5829679

fbshipit-source-id: 06995044b48fdadde02dece1624322a6e8b160bc