report merge base change when result set is empty
Summary:
In SCM aware mode, file changes and merge base changes are reported
to subscribers (https://facebook.github.io/watchman/docs/scm-query.html).
An optimization is to suppress subscription PDUs when the file list is empty.
This optimization needed to be further restricted to scenarios where the
merge base has not changed (and the file list is empty).

Reviewed By: wez

Differential Revision: D6077350

fbshipit-source-id: a69f874f52bd275a083af91c514ff2477cca2f85