Fix watchman's failing scm aware tests
Summary:
SCM aware tests were constructed to check that the file list was empty for scm
aware subscriptions. The file lists were empty because the user did updates
along the master branch. The problem appears to be related to the fact that the
state-leave command (which would be run at the conclusion of a checkout) does
not do a syncToNow(). This creates a race condition in the test's subsequent
call to flush-subscriptions (which does not defer notifications).

Reviewed By: wez

Differential Revision: D6101194

fbshipit-source-id: 7692b05b97f69a530cfb11541daad4b80c4d9973