Add utility method and tests
Summary:
This commit provides a utility template method to check if a map contains any
of a list of keys. The keys are provided using start/end pointers (iterator).
Included are a number of test cases to exercise the new code and validate the
implementation.

There is a need for these changes (in stacked commits).

Reviewed By: wez

Differential Revision: D6203377

fbshipit-source-id: 9afe0f0cf8a4d4a82e94009c53a500a67a159976