Defer sending subscription results when state changes observed
Summary:
TODO: I will add a new cli command and update tests prior to submitting
this change.

In SCM aware mode, subscription queries may be deferred, depending on
asserted states. This check is done before the query is run. Testing has
revealed scenarios where the subscription queries start running and before
they complete, there is a SCM operation that changes the merge base.
In these scenarios, watchman may over-report file changes. We can
prevent this over-reporting by determining if there were any state-change
operations interleaved with the subscription's query.

This change adds a state change counter to the root and increments it
when state changes occur. The state change counter is checked before
and after an scm state change query is run. Thus, we can determine if
there state changes occurred concurrent with query execution. In this
case, we discard the query results and restore the subscription state
(to the state before the query ran).

Reviewed By: wez

Differential Revision: D6165962

fbshipit-source-id: 02cb80e2cae51612cb3f91aafa2e33af10282822