cache getFilesChangedBetweenCommits when called for eden
Summary:
Buck currently performs a watchman query against each of the
configured cells sequentially.  When invoking buck after a commit or amend in
eden, watchman has to expand the list of changed files by running a mercurial
command.  This command can take several seconds to run, multiplied by the number
of configured cells.

This diff adds a cache so that we pay that same status cost just once.

Follow on diffs will work on eagerly computing this difference so that the
latency for buck is reduced.

Reviewed By: eamonnkent

Differential Revision: D6299488

fbshipit-source-id: 0baa2368eeea958e4c82946317caf2e26c5bca08