implement pidfile locking on win32 (sort of)
Summary:
I was debugging an integration issue with Nuclide on
windows earlier today and was surprised to see that we had
several instances running.  What was happening was that each
time I'd asked watchman to terminate, each of the subscriptions
established by Nuclide would get a disconnection error and
then independently use the watchman CLI to restart the server.

It turns out that I'd stubbed out locking the pidfile on windows
(with a fun comment that had me smiling rather than swearing at
my past self), and so it is time to actually implement locking.

Unix style pidfile locking is not easily implemented on windows as the locking
primitives are quite draconian; you can either exclusively lock the file for
write, denying all others all kinds of access, or lock it for shared access
denying everyone (including the lock owner) from writing.  There doesn't appear
to be a way to downgrade the exclusive lock to a shared lock which means that
writing the pid to the file is not possible without introducing a race
condition.  The file locking APIs are also rather awkward to use, requiring
a byte range and length both expressed as pairs of 32-bit values.

I've opted to sidestep that here and use the traditional win32 single
instance approach of using a named mutex.

There are some comments inline about the use of the Global namespace for
this also.

Because of this difference in approach it is not possible to report
the process id of the process that owns the lock on windows, but that
isn't strictly required anyway.

Reviewed By: ilya-klyuchnikov

Differential Revision: D6609479

fbshipit-source-id: d9d65e82dcfdf36520d147cb86cc89c67bec1c56