Support specification of benchmark iterations
Summary:
Add support for "bench_iterations" in query specification. This is done in a
back-ward compatible way. That is: if "bench" alone is specified,
query to be done 100 times. However, if "bench_iterations" is present and a
valid value, its value will be used.  Here is an example:

watchman -j --sockname=$WATCHMAN_SOCK <<-EOT
["query", "/data/users/ekent/opsfiles-hg", {
  "dedup_results": true,
  "bench": true,
  "bench_iterations": 6,
  "empty_on_fresh_instance": false,
  "since": "c:1515806033:3940221:3:2",
    "expression": [
      "name",
      ".xyz/abc",
      "wholename"
    ]
}]
EOT

Reviewed By: wez

Differential Revision: D6718099

fbshipit-source-id: 39ebe09ad5f5484a987b8f4847d2f1f1878debe0