watchman: periodically call clock on watched roots
Summary: This change adds causes Watchman to periodically run the clock command on its watched roots. It gets the list of roots by calling 'watchman watch-list'. It subsequently calls clock on each watched root. It will log the time that it took to run the clock command. This will provide an interesting metric for analysis.  For example, if the clock command is taking an excessively long time, there may be lock contention issues.

Reviewed By: wez

Differential Revision: D6804026

fbshipit-source-id: a08122c46a74426b1daa5d8af21df83ac1f1221f