watchman: add command to trigger jemalloc heap profile generation
Summary:
This is enabled only for fbcode builds internally at the
moment, because it easiest not to wrangle with other build systems.

This is only enabled for builds without ASAN enabled.  Usage is:

```
$ buck build watchman/... mode/dev-nosan
$ rm /tmp/wat
$ env MALLOC_CONF="prof:true,prof_prefix:/tmp/jeprof.out" \
         WATCHMAN_CONFIG_FILE=/dev/null \
         ./buck-out/gen/watchman/watchman -f \
         -U /tmp/wat -n -o /dev/stdout \
         --statefile=/tmp/watstate --pidfile=/tmp/watpid
```

In another window:

```
$ watchman -U /tmp/wat debug-prof-dump
{
    "version": "4.9.0",
    "prof.dump": "mallctl prof.dump returned: Success"
}
$ ls /tmp/jeprof.*.heap
/tmp/jeprof.out.2999124.0.m0.heap
$ mkdir -p ~/public
$ jeprof --gif ./buck-out/gen/watchman/watchman /tmp/jeprof.out.2999124.0.m0.heap > ~/public/wat.gif
Using local file ./buck-out/gen/watchman/watchman.
Using local file /tmp/jeprof.out.2999124.0.m0.heap.
Dropping nodes with <= 0.0 MB; edges with <= 0.0 abs(MB)
```

We deliberately don't allow passing the filename for the profile dump
through for security reasons; we don't want to cause watchman to write
to an arbitrary file in response to a client request.  It would be
nice to generate the name and return it in the response, but that
can be added later.

Reviewed By: eamonnkent

Differential Revision: D6879911

fbshipit-source-id: f2d736017554c0e5afb9542fe3e4d3ac1dfe98eb