fix trigger test flakyness/hang
Summary:
This was happening 100% of the time for me on my mac, so
it was a good time to fix it.

What seems to be happening is that an unrelated problem with the
PubSub stuff that drives the `log-level` unilateral stream gets
confused and never delivers us the logs of interest.  We're
tracking that internally as T26117870.

Secondly, the code that waited for the log files had its own internal
loop that was kept busy streaming log data from the above malfunction
and didn't respect the overall waitFor timeout.

Thirdly, because we have some funky signal handling stuff to safely
tear down the multi-threaded test runner, we'd swallow interrupt
signals and never break out.

The resolution is twofold:

* Add a helper to check and raise a KeyboardInterrupt at an opportune
  time in the code that is waiting.   It's worth following up to
  re-evaluate this interrupt handling stuff because it may not be
  needed any more; it may have been paving over the threads vs fork
  problems on python 2.7 and we dodge that bullet through other means
  today.

* Change the test to match against the log file rather than the log
  stream.

(This diff includes a couple of PEP8 fixes)

Reviewed By: simpkins

Differential Revision: D7003957

fbshipit-source-id: 70137d1c45feb148c4aa4f4de4d815fc703cdddf