detect when cookie dir has gone awol
Summary:
I got here because `btrfs` doesn't send `IN_DELETE_SELF` notifications
when a watched subvolume is unmounted.  Until that is addresses in the kernel
we tackle this from a different angle.  We periodically perform a synchronized
`clock` call against each of the watches.  When that happens and a btrfs volume
has been deleted then the cookie sync will error out because the directory
structure no longer exists.

Let's catch that case and generate a recrawl; the recrawl will discover the
removal and cancel the watch.

And while we're in here, let's also deal with just the vcs subdir going away.

Refs: https://github.com/facebook/watchman/issues/25
Refs: https://github.com/facebook/watchman/issues/501

Reviewed By: simpkins

Differential Revision: D7014364

fbshipit-source-id: 9acd20efa843563626b73c6f6e34c3787dd28a39