connect crawl errors to the watch failure reason
Summary:
This is a bit of a clean up in preparation for sharing
and knowing more about the reasons behind a failure of this
and cookie sync timeouts.

During crawling with inotify, if we hit the inotify system limits,
we have logic to set us in a poisoned state and fail new requests
until we are restarted.

An external user reported a problem where the `watch` project
command would appear to succeed when it hit the limits rather
than surface the error.

https://github.com/facebook/watchman/issues/525

The root of this problem is that we weren't populating the
`failure_reason` field to bubble up the error and that is the
core of this diff; with that set, the error is reported in
the result of the `watch` command.

While looking at this I noticed that we had a couple of stale
bits that needed cleanup:

- inotify.cpp had legacy code for setting the poison state itself,
  rather than letting the logic in warnerr.cpp handle it.
- the error_category stuff needed to be educated about the system
  limit errnos that the legacy code above was using
- the code in warnerr.cpp is now responsible for setting the
  `failure_reason` if it wasn't already set, and it now will take
  the poison reason if it is set (it's a global variable).
- When we encountered an error on the root of the watch we had
  a notion of whether the issue was transient or not.  I don't
  recall why that matters there, and all the errors that we can
  experience in practice are sufficient that we should bail
  out on watching that root, so I've removed the transient flag
  and revised the error messaging.

Reviewed By: simpkins

Differential Revision: D7063735

fbshipit-source-id: 1da4ad5f53b8fb9eb8daeb4c36ab953a1a75f0b5