work harder to surface sync errors more accurately
Summary:
This rearranges things a little to make it easier to propagate the
reason and context behind a cookie sync timeout; we now use an exception for
this so that we can include data about the recrawl state of the repo too.

While testing this I noticed that we would block for a long time even if we set
the sync_timeout to be super short and I traced this to the perf sampling code;
it was collecting the current clock information from the root, but this data is
locked during a recrawl and this meant that our wonderfully short timeout
wasn't being respected.

In practice, I've never needed to look at the current clock position in the
data that we sample to scuba so I've simply removed it to unblock things.  This
should help to avoid situations like this one that was reported in the group:

https://fb.facebook.com/groups/scm/permalink/1572802659436015

Reviewed By: simpkins

Differential Revision: D7075600

fbshipit-source-id: 2b70a814a55d7b5d5d0d60d204ee0d6cc825f3c6