improve error reporting if site spawner fails
Summary:
The stock behavior of watchman is to `daemonize` the server process if
needed, but some sites have more elaborate requirements; the site spawner
configuration allows the local administrator to deploy a policy for customizing
how watchman is launched.  We use this at FB to managed watchman instances for
users across our dev fleet via chef and systemd.

There were a couple of issues with how we surfaced errors:

1. In one possible spawn error case we were logging an error
   at `ERR` level instead of `FATAL`.  This meant that we'd
   return to the caller (the watchman CLI) without any hope
   of the server starting up and there is no fallback to
   daemonize when configured this way, so the CLI would sit
   and wait until its timeout was hit.

2. We'd attempt to log the site spawn failures to the log
   file rather than the current process stdio streams.  This
   made it rather difficult to diagnose problems with the
   site spawner.

This diff fixes up the missing `FATAL` and removes the stdio redirection so
that we report spawn errors more directly to the user.

Reviewed By: simpkins

Differential Revision: D7076476

fbshipit-source-id: 3f001e56fabea4c5c630d6110e19a671c9efbfb1