speculative de-flake of test_wm_wait.py
Summary:
This seems to fail reasonably frequently on appveyor and
doesn't show any useful context, so let's try to make it less flaky.

This is basically a rewrite to make this whole thing simpler.  It turned
out that there was a race waiting for the root to be watched, so this
could perhaps have been a minimal diff that just added that check,
but since it seemed so complex it felt worthwhile to simplify.

```
FAIL test_wm_wait.TestWatchmanWaitLocalBser.test_wait (5.431s)
Traceback (most recent call last):
  File "C:\Python27\lib\unittest\case.py", line 329, in run
    testMethod()
  File "C:\projects\watchman\tests\integration\test_wm_wait.py", line 60, in test_wait
    wi, ['a/bar', 'a/foo', 'a'], msg="wait, create dir")
  File "C:\projects\watchman\tests\integration\test_wm_wait.py", line 37, in assertWaitedFileList
    self.assertFileListsEqual(res, expected, message=msg)
  File "C:\projects\watchman\tests\integration\WatchmanTestCase.py", line 320, in assertFileListsEqual
    self.assertItemsEqual(list1, list2, message)
  File "C:\Python27\lib\unittest\case.py", line 901, in assertItemsEqual
    self.fail(msg)
  File "C:\Python27\lib\unittest\case.py", line 410, in fail
    raise self.failureException(msg)
AssertionError: wait, create dir
FAIL test_wm_wait.TestWatchmanWaitLocalBser.test_wait (5.437s) (1 attempts)
```

Reviewed By: simpkins

Differential Revision: D7077906

fbshipit-source-id: abfdf2b1a73b3f446b29802516cf4376c283f934