fix race condition in fsevents watcher when removing root
Summary:
`test_remove.py` was flaking out on travis due to this issue.
From the logs it's clear that we don't recognize that the root has
been removed.

Add an additional check for the root being removed rather than
relying on the other signals from fsevents about unmounting and
other changes to the root.

Differential Revision: D7160741

fbshipit-source-id: e125596d32152ebe441ca170808d7b2ac967b344