fix potential dangling reference
Summary:
I saw a SEGV stack trace in this function in the coredumper dataset.

Ideally we'd thread a shared_ptr through here, but the fieldlist is typically
a very small vector of a couple of pointers, and this async render code path
is used rarely (only when computing content sha1s) so the relative cost should
be neglible.

Reviewed By: simpkins

Differential Revision: D7191799

fbshipit-source-id: b9137a11d7c903f18e54ef88cb16f6a66130f50f