remove stack based w_string hack
Summary:
the hot paths now use the code that uses string_piece
for keys.  There was one code path in the json bits that was
using it, but it is not in the hot path: we would use it only
during parsing of commands so relatively speaking it is cheap
to do without.

We can always add overloads that take w_String for that if
needed.

Reviewed By: chadaustin

Differential Revision: D7209342

fbshipit-source-id: 7403a1bbe750a06dd43123b18f970c2d995b6dd9