make sanity check thread persistent
Summary:
When the sanity checking stuff was first added it
was called every minute to validate that the unix domain
socket still resolved to ourselves.  It was added as a detached
thread that was short lived and that terminated as soon as
the check was complete.

In D6804026 the scope of the checks was expanded and it is now
possible for systems under severe IO load to take longer than
1 minute to complete the sanity checks.  When that happens
we can end up stacking up instances of the sanity checking
thread and magnifying the number of clients that are blocked
waiting for progress.

To avoid that, this diff makes the sanity checking thread a
long lived thread that is responsible for timing and dispatching
the checks.   That means that we will never have multiple competing
sanity checking threads and a maximum of 1 outstanding sanity
checking session at a time.

Reviewed By: chadaustin

Differential Revision: D7220354

fbshipit-source-id: 83420c7ddeb9188518d046948b0dac645da6f156