fixup memory corruption issue when parsing path generator
Summary:
This was reported as an intermittent failure.  We didn't have
integration test coverage for queries of this form and thus missed
it in the recent sweep to replace some old legacy string handling.

I've added an integration test that reproduces the issue in ASAN mode
and fixed it up by migrating away from the older `json_unpack` API.

Reviewed By: simpkins

Differential Revision: D7307742

fbshipit-source-id: 9df745612ce6bd151b465614cbf3cd65cbec09b7