state-(enter|leave) are now non-blocking
Summary:
We're seeing some slowdowns in hg update operations; the blockage
appears to be waiting on state-enter and state-leave commands issue by the
fsmonitor extension.

This diff builds on the prior diff and makes use of the more generalized cookie
sync mechanism to arrange for the state assertion action to take place once the
cookie has been observed by the IO thread.

Note that this breaks state-enter/state-leave for the eden watcher, which
currently does not process cookies, however, the fsmonitor extension is not
even enabled for eden, so this is sort of a moot point; we'll need to resolve
that in a different way.

This diff complicates things a little; previously, we would not update the
client->states map until after synchronization had occurred.  This made it very
simple to understand who had made a state assertion and allowed for some strong
sanity checks.

With this diff, the visibility of the assertion is deferred until the cookie is
processed; a client may therefore be pending an enter or leave.  In the case
where mercurial is processing a series of updates (as part of a rebase), it is
possible that it may have issued a series of enter/leave events and for the
visibility of those to be queued up.  For example, there may be a pending
`[enter, leave, enter]` at the point where fsmonitor wants to issue the closing
`leave`.

Handling this is reasonably straight forward; when processing a leave request,
we look for the item with the smallest id number and check based on that.  The
id numbers are assigned in increasing values, so the smaller numbers imply the
earlier event.

Reviewed By: simpkins

Differential Revision: D5118348

fbshipit-source-id: b6f8aca6a25e61b1daf2d4e80fd6eea7544174a4