use otime rather than timestamps for symlink cache
While timestamps seemed to work reliably on our internal CI, they're
pretty flaky on the open source CI and also on my laptop.

Switch the cache key to use the `otime` field instead; this holds
the watchman clock value from the last time we observed a change
for a given file node and isn't subject to timestamp granularity
or clock drift concerns.

Test Plan: test_symlink.py would pretty consistently (about 80% of the
time) repro a failure when using bser over a a unix socket.  With
this change we're pretty solid.

Let's see what the OSS CI makes of it.