terminate process without running destructors on FATAL
Summary:
For better or worse... I've seen this trace in the wild,
and its been lingering for 4 days!  Strangely, the process is almost
fully functional, except for flushing subscriptions.

I suspect that we have an uncaught exception in the client thread because the
logs look like this; these indicate that the terminating thread was a client
thread.  We can see that it handled a query 2 seconds prior to the fault:

```
2018-03-17T04:52:48,510: [client=0x7f7f39d3e1d0:stm=0x7f7f39c75840:pid=1819024] PERF: {"system_time": 0.081978999999999996, "user_time": 0.0055560000000000002, "elapsed_time": 1.2319009999999999, "description": "que
ry_execute", "meta": {"query_execute": {"query": {"expression": ["allof", ["not", ["anyof", ["dirname", ".hg"], ["dirname", ".git"], ["dirname", ".svn"]]]], "since": {"drop": ["hg.update"], "scm": {"mergebase-with":
 "master"}}, "relative_root": "", "empty_on_fresh_instance": true, "fields": ["name"]}, "fresh_instance": false, "num_deduped": 0, "num_results": 1, "num_walked": 11}, "root": {"watcher": "inotify", "path": "/data/u
sers/akramer/www-hg", "case_sensitive": true, "recrawl_count": 0}}, "start_time": 1521287567.278764, "pid": 3700551, "version": "4.9.0"}
2018-03-17T04:52:50,116: [client=0x7f7f39d3e1d0:stm=0x7f7f39c75840:pid=1819024] std::terminate was called
Fatal error detected at:
/usr/local/bin/watchman() [0x411507]
/usr/local/bin/watchman() [0x6c12f6]
/usr/local/bin/watchman() [0x443620]
/usr/local/bin/watchman() [0x6c2525]
/usr/local/fbcode/gcc-5-glibc-2.23/lib/libstdc++.so.6(+0x8cd95) [0x7f7f3b8ded95]
/usr/local/fbcode/gcc-5-glibc-2.23/lib/libstdc++.so.6(_ZSt9terminatev+0x10) [0x7f7f3b8dede0]
/usr/local/fbcode/gcc-5-glibc-2.23/lib/libstdc++.so.6(+0xc7227) [0x7f7f3b919227]
/usr/local/fbcode/gcc-5-glibc-2.23/lib/libpthread.so.0(+0x77a8) [0x7f7f3b0157a8]
/usr/local/fbcode/gcc-5-glibc-2.23/lib/libc.so.6(clone+0x6c) [0x7f7f3a931a7c]
```

There was not a lot of useful additional information available from the process
(I captured a core file and poked around).  Here's the trace of the stuck
thread.  None of the other threads are waiting on locks; all threads appear to
be waiting for requests and operating reasonably; other requests for the same
root will succeed and so on.

```
Thread 75 (Thread 0x7f7f1ab7e700 (LWP 1819033)):
```

I've expanded our termination handler to attempt to print out the current
exception.  This seems to be present for the obvious cases; the trace doesn't
provide a huge amount of context but at least we can divine the thread name
from the existing logging, and can also resolve to the thread function when
attaching to the gdb core file.

```
2018-03-22T09:52:05,164: [perflog] std::terminate was called. Exception: boop.
Fatal error detected at:
./buck-out/gen/watchman/watchman() [0x69ccd4]
./buck-out/gen/watchman/watchman() [0x69e7fc]
./buck-out/gen/watchman/watchman() [0x443ee0]
./buck-out/gen/watchman/watchman() [0x6a062d]
./buck-out/gen/watchman/watchman() [0x6a07f8]
/usr/local/fbcode/gcc-5-glibc-2.23/lib/libstdc++.so.6(+0x8cd95) [0x7fc441cabd95]
/usr/local/fbcode/gcc-5-glibc-2.23/lib/libstdc++.so.6(+0x8b9d8) [0x7fc441caa9d8]
/usr/local/fbcode/gcc-5-glibc-2.23/lib/libstdc++.so.6(__gxx_personality_v0+0x387) [0x7fc441cab357]
/usr/local/fbcode/gcc-5-glibc-2.23/lib/libgcc_s.so.1(+0xff0c) [0x7fc4404cdf0c]
/usr/local/fbcode/gcc-5-glibc-2.23/lib/libgcc_s.so.1(_Unwind_RaiseException+0x1ac) [0x7fc4404ce44c]
/usr/local/fbcode/gcc-5-glibc-2.23/lib/libstdc++.so.6(__cxa_throw+0x63) [0x7fc441cac053]
./buck-out/gen/watchman/watchman() [0x488dad]
/usr/local/fbcode/gcc-5-glibc-2.23/lib/libstdc++.so.6(+0xc7170) [0x7fc441ce6170]
/usr/local/fbcode/gcc-5-glibc-2.23/lib/libpthread.so.0(+0x77a8) [0x7fc4413e27a8]
/usr/local/fbcode/gcc-5-glibc-2.23/lib/libc.so.6(clone+0x6c) [0x7fc440cfea7c]
[1]    1237117 abort (core dumped)  ./buck-out/gen/watchman/watchman --foreground --logfile=/dev/stdout
```

I also cleaned up the handling of the `send_ok` variable.  It appeared to be
possible for this to get set but then nothing would ever terminate the main
loop.  That code was added back before much of the supporting bits had been
converted to C++.

I've changed the `exit()` call to `_exit()` to avoid calling `atexit`
handlers and destructors.  We used to `abort()` here, but there are a number of
cases where we don't want that to generate a core file so it was recently
changed to exit.   I've added an explicit `ABORT` error level to accomodate
this class of problem.

Reviewed By: simpkins

Differential Revision: D7354472

fbshipit-source-id: 846f93b568beb1e7b00dc91f02e71c40678c86b6