static_assert json_int_t vs time_t
Summary:
It is important that we're able to represent the full range
of a `time_t` in a `json_int_t`.  This `static_assert` spells that
out.

This is important to check because a later change in this cmake stack
broke detection of `strtoll` for jansson on Windows, and the presence of that
function is what enables 64-bit integer types.

`time_t` is 64-bit by default on Windows.

Reviewed By: chadaustin

Differential Revision: D7366201

fbshipit-source-id: 195b5387d37493a56fc33058f722285d95f80b2c