win32: ERROR_BROKEN_PIPE -> EOF on read
Summary:
while working on the cmake-ification on Windows, I ran
into a strange error where ChildProcess would always cause the
watchman process to terminate in the trigger tests.  I traced
part of that to problems with assumptions around `echo.exe` being
available; in the case that the system doesn't resolve `echo.exe` the pipe to the child process is broken.

Windows will return ERROR_BROKEN_PIPE for both reads and writes,
whereas posix systems yield EOF on read and EPIPE (assuming that
that SIGPIPE is disabled) on write.

Our default mapping is performed in winbuild/errmap.cpp.

In the integration test scenario, the threaded pipe reading code
will translate EPIPE into an exception and bubble it up.

For reads only we want to map this to EOF.

Reviewed By: simpkins

Differential Revision: D7366197

fbshipit-source-id: dd9bd66c0c714755f672ec0379abb967ae5afb50