normalize separators in relative_root
Summary:
On windows, it is possible for the client to pass in a path
that has a mix of / and \ and that is effectively invalid on windows.
Our policy is to perform this sort of translation at the edges, so
this is the right place to do this.

Note that we were already canonicalizing the path; our canon operation
is really just removing a sequence of trailing slashes and doesn't mutate
the string in any other way.

Reviewed By: ilya-klyuchnikov

Differential Revision: D7758085

fbshipit-source-id: feef7c1d1ce9df6ccd0a43231dc96c7e68a31083