Update Watchman to use Eden's new globFiles() API.
Summary:
D7748870 introduced `globFiles()`, which improves over `glob()` by taking an
`includeDotfiles` parameter. Prior to this revision, the Eden watcher in
Watchman would ignore the value of `includeDotfiles` specified by the caller.
With this change, `includeDotfiles` is honored by forwarding it to `globFiles()`.

Because we do not have a reliable way to tell whether the Eden Thrift client we
are talking to has been updated to support `globFiles()`, we try `globFiles()`
first and then fall back to `glob()` if we get an error. Admittedly in the
fallback case, `includeDotfiles=false` will not be honored. (We could
post-process the list to honor it properly, but instead we'll focus on
updating the old Eden instances so we can delete this codepath.)

Reviewed By: simpkins

Differential Revision: D7849540

fbshipit-source-id: 4ea7578fc29616c9cd14c19dd63c9effb8bb7c92