prepare for more folly-dependent code
Summary:
kcoons has a stack of diffs that pull in some facebook internal
libraries.  These use singletons to initialize things early on, and this
impacts the start up time of the watchman binary and triggers some flakiness
in the `test_sock_perms.py`; it doesn't wait long enough for the binary to
emit the error message and thus fails.

This diff addresses the test flakiness by raising the timeout in the test;
note that this increases the minimum latency for the test, so we don't
want to raise this all the way up to 10 seconds.  Ideally we'd poll
for the output but we don't have a good way to connect our `waitFor`
helper to the instance startup code today.

I've also moved the magic glue for completing the folly singleton
initialization to main; this is basically where it was on the
very first iteration of D4725016.  It makes more sense for it to be
centralized now that we have both eden and the saved state modules
pulling in folly deps.

In the longer term we're going to take a full dependency on folly
and use folly::init, but for now we minimize the surface area.

Reviewed By: strager

Differential Revision: D8758271

fbshipit-source-id: c66feb35b5d567b29ba8ca1a26a595fcf88c3d82