make FileResult methods return Optional values
Summary:
Allow some of the accessors in `FileResult` to return optional
values to indicate that the value is not currently known.

A side effect of these methods returning an empty optional value
is that they will set the neededProperties_ member with an appropriate
value so that a subsequent batchFetchProperties() call will obtain
the require data.

None of the FileResult implementations return a empty value for
these today, so this diff should not change functionality.

Reviewed By: strager

Differential Revision: D8584080

fbshipit-source-id: 47373b7e4fc71bfd421e646004e48f41251df8b8