remove data dep on exists/type for sha1 result rendering
Summary:
We would previously request additional data to decide whether
to request the content sha1, which meant fetching more data than we
strictly need.  For example, in eden, we shouldn't need to stat a
file in order to retrieve its sha1.

This diff moves the checks that were previously in the rendering code
and pushes them into the FileResult methods themselves, allowing
the implementation to decide how best to proceed.

For InMemoryView and EdenFileResult the error conditions raise
system errors that we can now catch in the render code so that
the existing tests have preserved their behavior.

Reviewed By: strager

Differential Revision: D8651888

fbshipit-source-id: d7f70d9bdc27e3606d22a48e8ae17e8dcf7e38bf