fixup batch loading for content.sha1hex
Summary:
The main problem here was the use of `baseName()` in the loop
when computing the file names to examine for sha1hex and readlink.

This had the effect of concatenating the basename from the first file
in the batch with the dirnames of the subsequent entries, generating
nonsense filenames for the majority of results.

A secondary issue is that these bogus filename generated error results
which we considered to be equivalent to "need data" due to the use
of `hasValue` rather than testing whether the `Result` type was
`empty`.

The readlink code had a similar potential issue with the error case.

Reviewed By: chadaustin

Differential Revision: D9073833

fbshipit-source-id: d9ad330297b40593f9a605c1d015975ba96f0205