InMemoryView::batchFetchProperties must always join futures
Summary:
We've seen a crash occur in the case that the LRUCache is saturated
with requests.  When this happens, the cache code will throw an exception and
this would leave the set of futures spawned unjoined.  Since the continuation
of those futures doesn't (and cannot!) take owning references to the data it
needs to manipulate, it was possible for them to generate a number of invalid
memory accesses.

This diff uses SCOPE_EXIT to ensure that we always wait for the futures,
even in the exception case.

Reviewed By: chadaustin

Differential Revision: D9093692

fbshipit-source-id: 529572a543d6666c228833eb74440c8c0240029a