allow making progress even when content_hash_max_items is low
Summary:
The prior diff highlighted a crash bug in this scenario.  With the crash
addressed we can now look at the ergonomics in this situation.  The purpose of
`content_hash_max_items` is to set an upper bound on the long term memory usage
of the watchman server, so it is reasonable to allow bursting above that threshold
while satisfying the `getter` future associated with the cache.

This diff removes the exception throwing from `evictOne` and instead returns a boolean.

We now check for eviction at the start of a cache request and also at the end.  At
the end we will try to bring the usage below the configured limits rather than just
trying to remove a single item.

Reviewed By: chadaustin

Differential Revision: D9093948

fbshipit-source-id: 44815525ffaff2e967028ba69f3639e6d54ae918