Read saved state info from watchman clock
Summary:
Add a saved state JSON object to the fat (now obese) clock
in watchman, as a sub-component of the scm part of the clock.

The saved-state segment includes the storage type ("local" and "manifold"
are currently supported), a json_ref key that will be whatever type
the specific storage type requries, and a revision that indicates what
source control commit the saved state is associated with in the response,
or an empty string if no saved state was available.

The "saved-state" key is optional, but if the "saved-state" key is
specified then both its "storage" and its "key" must be specified. The
"revision" field is optional in the query. The only thing that will be
done with it is that it will be returned if the mergebase does not change.

I tried putting this data in the clock but outside the "scm" component,
outside the clock entirely, and finally here inside the "scm" component
of the clock. I liked this best becuase:

1. Specifying saved state options without an scm-aware query makes
no sense.

2. The saved-state revision modifies the "files" returned, and the set of
files returned is tied to the clock returned, so it seemed appropriate
that the clock would be different if the files were relative to a saved
state than if they were not. If the saved state info were outside the
clock, then the clock could be identical with two different sets of
changed files depending on whether a saved state was available or not,
and that seemed unintuitive.

Reviewed By: wez

Differential Revision: D8306327

fbshipit-source-id: 6cc898ed635d0f5e127dcb0003210106e4546b47