Update watchman result with optional saved state
Summary:
Update the result of the watchman query with saved state info
if the query includes a saved state specification in the clock.

Update the query, since, and subscribe commands to update the response
with an info blob that contains storage-type specific information about
how to access the actual saved state, if one is available.

The saved state info will only be available if the mergebase has changed,
otherwise it is not needed and we do not compute it.

If the mergebase did change, the client can tell whether a saved state
is available or not via the saved state revision in the clock, which
will be an empty string if the mergebase has changed yet no saved state
was available. In that case the info blob will still be present, but it
will contain error information.

This saved state info blob is in the main response along side the "files"
returned because it's part of the response only, and it, when combined
with the "files" field, determines what the new state of the world is
for the client.

For now, the fallback if no saved state is available is to the scm-aware
behavior, where the client gets back the files changed since the
mergebase with trunk. This is the most logical fallback given the API
(and the simplest to implement), but it's totally useless to the client
most likely. If no saved state is available, then the client must fall
back to all changes since the last clock. As a follow up, an option
to control the fallback might make sense, so that clients don't need
to throw the response away and perform a subsequent query to get all
changed files if the saved state lookup fails.

Reviewed By: wez

Differential Revision: D8306759

fbshipit-source-id: 87ac7809f4a41d844feaea148bb171c91cb7c153