fix importing facebook.eden in eden integration tests
Summary:
When the facebook specific integration tests were added
under `watchman/tests/integration/facebook`, we broke the ability
to resolve the `facebook.eden` thrift types used by the eden
thrift client and caused the eden integration tests to skip.

This grubs around with the python path to allow the tests to work again.

Reviewed By: chadaustin

Differential Revision: D9307105

fbshipit-source-id: 7732bf0940f98844b0c306a400920096efb14223