fix unterminated render loop when rendering a deleted file in eden
Summary:
This issue manifested most often with the subscriptions used by Nuclide.
In the scenario I observed, the `.hg/merge` file had been removed and I saw that
watchman was repeatedly querying for this file.

Since its query had requested the `mode` of the file, we were calling into the
property accessor that was gated on whether `stat_.has_value()`.  The code that
processed the `getFileInformation` results was setting `exists_` but then clearing
out the `stat_` optional, so we'd never believe that we had all the data needed
to render the results to json, and just spin forever.

Reviewed By: chadaustin

Differential Revision: D9309012

fbshipit-source-id: 12efd1dbc719c6b95f2c6563299012c54937ee68