pywatchman: fixup error handling on win64
Summary:
While this code worked on 32-bit windows, it would misbehave
on 64-bit windows.  The main problem was that the invalid handle value
would never match; this caused most of the tests in our test suite
to incorrectly diagnose an error at connect, in most cases causing the
warm up code not to wait long enough for the server to come up.

I've improved some general error capturing cleanliness and reporting
as part of tracking this down.

Refs: https://github.com/facebook/watchman/pull/636

Reviewed By: simpkins

Differential Revision: D9396867

fbshipit-source-id: f09f77bd4ec7b606cadbc087f9ad0d3ac46699c6