check server liveness between tests
Summary:
Sadly, a source of flakiness in the windows CI is the server
having problems in between tests.  When this happens, the server
associated with that worker thread remains dead for the rest of the
test suite.

With the move to folly, we have 10-15 minutes of latency just fetching
and building the folly deps, so this kind of flakiness is undesirable.

This diff adds a liveness check in between running each test; if the
server appears to be dead, it cleans up the old one and spawns a new
one.

This allows the test suite to complete.

It does mask some server instability.  This is a reasonable trade:
it is impossible to diagnose those problems on appveyor, and they
did not reproduce locally.  I don't feel great about this.

Refs: https://github.com/facebook/watchman/pull/636

Reviewed By: simpkins

Differential Revision: D9397010

fbshipit-source-id: 00e3f949e0b34e814e3765bb08e197df925a9c45