upload appveyor artifacts to logs7.zip when --win7 is used
Summary:
Since introducing folly deps increases the build duration to
20 minutes per job, it is desirable to eliminate one of the jobs from
the matrix.   Currently the appveyor build matrix has two rows; one
for the base build and another with an environmental variable set
to run the tests in a mode to enable win7 code paths.

A later diff in this stack will replace the row in the matrix with
just adding a second test run with the `--win7` flag passed to
runtests.

When that flag is used, make sure we capture the logs from that run
in a separate zip file.

Refs: https://github.com/facebook/watchman/pull/636

Reviewed By: simpkins

Differential Revision: D9397011

fbshipit-source-id: ee932aebe8736690458b14ef285c75a4cb1fff08