pywatchman: detect use of a client after a fork
Summary:
We sometimes see what looks like invalid state-enter/leave
transitions in the mercurial integration.  One theory is that we're
inheriting a live connection across a fork boundary when multiprocessing,
and thus inheriting associated states.

This diff records the pid at the time that we establish a connection
and sanity checks that it matches the current process pid before each
read or write call.   We already centralized checking whether we
needed to open a connection in this place, so thankfully there's
just a single place to insert this logic.

One could argue that we could automatically attempt to open a new
connection here.  It feels "better" to be more determistic and always
report an error in this case.   That way the caller has the option
to catch and handle the error differently.

Reviewed By: quark-zju

Differential Revision: D9439758

fbshipit-source-id: 77eaf911a53e7acf50af9760f16adf9e61acb783