Return byte[]s for non-UTF-8 file paths.
The watchman docs note that, due to file system APIs being
encoding-less, we can't assume the bser strings are UTF-8.

This changes the bser string deserializer to fallback to returning
a byte[] when UTF-8 deserialization fails.

This lets client to a cast/instanceof check on String or byte[]
and decide how they want to handle the value.