avoid a couple of Eden thrift calls in subscription processing
Summary:
A couple of code paths call getMostRecentRootNumberAndTickValue()
a couple of times.  We can do this once and pass it down to the helper
function.   There's no need to make the second call; in this context it
is fine to use the value we computed a few lines prior.

(Note: this ignores all push blocking failures!)

Reviewed By: strager

Differential Revision: D9652041

fbshipit-source-id: 071d53b6086c27b41921f56931e37b5ea19def6f