catch some classes of error in processSubscriptions
Summary:
I noticed this issue while looking into an issue reported in
one of our internal groups.   The user had an issue with their Eden
daemon that was causing the subscription processing to trip up:

```
2018-09-04T16:09:49,111: [client=0x7f78ecaafa10:stm=0x7f78ecb83940:pid=1960898]
std::terminate was called. Exception: readlink(/data/users/USER/REPO/.eden/socket) failed: Stale file handle.
```

The stack trace logged was in the subscription dispatching code.
From inspection and from the error message, this must have been thrown while
setting up the eden thrift client.

This condition causes the program to terminate because the function
higher up in the call stack is marked `noexcept`.

In this case we don't want to halt the program, we just want to stop
trying to work on the broken repo, so let's catch this error and
cancel the watch.

(Note: this ignores all push blocking failures!)

Reviewed By: chadaustin, strager

Differential Revision: D9652040

fbshipit-source-id: a54e48aeb450b64bc6251e1b19f50c9222e34828