add basic retry on ESTALE when obtaining an eden client
Summary:
This is the lowest complexity approach for dealing with an unexpected ESTALE
response; a small retry loop in the event that we encounter an ESTALE error.

In practice we see ESTALE manifest in the `readlink` call when resolving the socket,
but it is also possible for it to bubble up for any syscall interacting with that
path.

We've been discussing whether we should cache connections or the results of readlink,
but managing invalidation and liveness checking is relatively complex.  For now
we need to have a quicker solution to the ESTALE problem that we can deploy to avoid
annoying interruptions in eg: buck build.

I've made this loop a few times with backoff because it seems like this condition
can be sticky for a short period of time, and a single retry didn't guarantee
success.

Reviewed By: strager

Differential Revision: D9951798

fbshipit-source-id: d6c0b290af0f4e5e2b12dab49946d256c7a300da