watchman: fix folly build on macos
TL;DR: cannot build with cwd = folly/folly because String.h is
found instead of string.h and this breaks all the system headers.

This diff generates a separate build dir for each cmake project
and builds from there.

Additionally, the shared library build of folly failed for me
due to a duplicate symbol error, so turn that off.