never attempt to monitor FUSE filesystems with inotify
Summary:
Update the InotifyWatcher to explicitly disable itself if the root is on a
FUSE filesystem.

It looks like there are some recent-ish proposals for ways to add inotify
support to FUSE (https://github.com/libfuse/libfuse/wiki/Fsnotify-and-FUSE).
However nothing appears to be currently implemented.  If FUSE ever does get
inotify support we can remove this check at that time.

Currently it can cause a bad experience for users if watchman does incorrectly
try to watch a FUSE filesystem using inotify: Watchman will claim to be
watching the filesystem, but it will not actually detect file modifications.

Reviewed By: wez

Differential Revision: D13336817

fbshipit-source-id: 92c69a7b3afa6541d8d03c825cfcbb07a8eb3ba5