port getdeps.py from eden
Summary:
This script is responsible for obtaining the build deps
for watchman.  It knows how to fetch the appropriate packages
on ubuntu (via apt), and windows (via vcpkg).

Some deps are fetched and built "manually" because they are not
packaged, or because they move too quickly to package (eg: folly).

Refs: https://github.com/facebook/watchman/pull/636

Reviewed By: strager

Differential Revision: D9397048

fbshipit-source-id: 71dc87eabf08b0b2093a693da51b1fedc02fe7ed