Link watchman against folly (#636)
Summary:
In theory this is the minimum needed to unconditionally link to folly.
It requires pulling in a fairly big set of dependencies that can be hard to manage,
so the `getdeps,py` script has been expanded with a little bit of smarts.

This diff started life as a pull request on github a few months ago and once it passed
the CI there we imported it into the FB internal machinery.  Since it has been idle
for a while it is probable that now that I've got it building on the FB internal CI
some additional changes will be required to make it build successfully on the OSS CI again.

The main code changes in this diff are to replace some of our windows portability bits with those in folly and to unconditionally initialize folly.

The bulk of this diff is messing around with build dependencies and environment.

Reviewed By: simpkins

Differential Revision: D9396698

fbshipit-source-id: aaa98214c6578b52df979725589c57ee04da4255