probe for dll deps (#677)
Summary:
For our windows packaging, now that we have dynamic deps via folly, it is important to collect the non-system deps together for distribution.

This diff adds a script to collect them into the `.zip` archive produced by appveyor, and in the FB internal CI, runs the same script but does not yet change the internal package to include those deps.

Pull Request resolved: https://github.com/facebook/watchman/pull/677

Reviewed By: chadaustin

Differential Revision: D13869172

Pulled By: wez

fbshipit-source-id: 5ca42d292fff0b7e7803a211ebd9b98d216029e3