watchman: fix local var type, avoid badness for config gets
Summary:
Found these while code modding in the next diff; these were using an
explicit json_t* type for the local variable which is implicitly convertible
from the json_ref returned by the functions they call, but is potentially bad
because the ref is released before the rest of the function is done consuming
it.  In practice I think this is fine because the config is generally not
mutated.

Reviewed By: chadaustin

Differential Revision: D14102883

fbshipit-source-id: ba8ed39edb616e259f2afa6ad5eeb4ca0bb82e54