watchman: add dtype concept to FileResult
Summary:
introduce the `dtype` method with a default implementation
that calls `FileInformation::dtype`.   Make the `type` field renderer
and `type` query term aware of `dtype`, which makes them potentially
very slightly faster as they can now be executed using a switch
statement rather than a series of conditionals on posix systems.

This should not change behavior in an observable way.

Reviewed By: strager

Differential Revision: D14111365

fbshipit-source-id: 08e71fe032f0536daf4084d3bbad2da5548f8530