watchman: add script to bootstrap a vcpkg install for windows deps
Summary:
This diff sets up a script that can re-create the vcpkg installation needed to build the dependencies for watchman on windows.   The intent is to share this with the eden build.

The bootstrap creates a zip file containing a vcpkg installation and then uses the `lfs.py` tool to upload that archive and capture its contents hash.

As a proof of concept, this diff includes that hash and the changes needed to the watchman lego windows build to stop using the fbpkg version of these deps (T40874235)

Reviewed By: pkaush

Differential Revision: D13954452

fbshipit-source-id: 606b381c3b447ffbbb17e71ac1886e2212295340