Clean up test edenfs if Watchman fails to start
Summary:
In the past, WatchmanEdenTestCase's setUp method started edenfs then starts Watchman. If starting Watchman fails, edenfs wasn't cleaned up. This is because the unittest module does not call tearDown if setUp fails.

D14294029 fixed this issue by starting Watchman before starting edenfs. However, if more code is added to WatchmanEdenTestCase.setUp and that code raises an exception, edenfs won't be cleaned up.

Instead of cleaning up edenfs in tearDown, clean it up using addCleanup. Functions registered with addCleanup are called if setUp fails.

tearDown is called before addCleanup callbacks are called. edenfs cleanup requires that the Watchman server and client are still alive. If we clean up edenfs in an addCleanup callback, we are forced to clean up the Watchman server and client using addCleanup too.

For tests using WatchmanTestCase but not WatchmanEdenTestCase, this diff should not change behavior.

Reviewed By: chadaustin

Differential Revision: D14268738

fbshipit-source-id: e02bceb4eb3071323cc01d21097767dfb57d1bfd