watchman: disable test_scm for ubuntu hosts
Summary:
We were already doing this but only for jobs spawned directly via NBTD.

The fbcode_builder spawned jobs don't have a good way to pass down the
environment variable that was used to skip this test, so this diff changes the
test to look to see if we're on an ubuntu host where this test is most
strongly flakey rather than looking at an env var.

We should spend some time to root cause this, but for now, this should reduce noise.

Reviewed By: mzlee

Differential Revision: D14383916

fbshipit-source-id: 5403c713c515db49ab93824f16523a503550e621