Add WatchmanClient::query
Summary:
I want to be able to run a single one-off query on a `WatchPathPtr`, to get all changes since a given clock.

`run` allows me to run queries, but requires me to pass the relative path to the repo via a string. And if I used `watch` and got a `WatchPathPtr`, I can't access the relative path inside the `WatchPathPtr` from outside the client.

`subscribe` takes a `WatchPathPtr`, but it creates a whole subscription with a whole callback.

I guess I could just use `subscribe` and then unsubscribe after a single result comes back, but that feels a little heavyweight.

Ultimately it feels like there's a gap in the API, hence this diff.

Reviewed By: strager

Differential Revision: D14382325

fbshipit-source-id: 7f5ff3b3b183d43ce74f0c931b4343cbd30e5f2f