Make WatchmanClient correctly set WatchPathPtr::relativeRoot_
Summary:
When you run `WatchmanClient::watch` on a subdirectory of a root that's already watched, you get back JSON with two fields:

```
{
  "watch": "/data/users/jhemphill/fbsource",
  "relative_path": "fbcode",
}
```

`WatchmanClient` then uses this data to create a `WatchPathPtr` which encapsulates a watched root. This allows clients to pretend that *their* root is being watched, even though behind the scenes Watchman may be watching a superdirectory of your root.

It looks like the WatchmanClient was checking a field called `"relativePath"` instead of `"relative_path"`, though. This diff fixes that typo and augments the test.

Reviewed By: strager

Differential Revision: D14392897

fbshipit-source-id: c1733b67b3644ce5e5c6fc775f1c7cba1618fa2c