watchman: assume we were built with cmake in childproc test
Summary:
there are no guarantees that there is an `echo.exe` in
the path on win32.  Previously, we would detect that we were built
using cmake by looking for a preprocessor defined symbol.  That
symbol is only present when generating a vcxproj and isn't available
when using the ninja generator.

Since the build requires cmake let's just assume that we have it
available in the path.

Reviewed By: strager

Differential Revision: D14702583

fbshipit-source-id: 6adcaa4593397f0ceebf78cd5cafa28cc8bd5b09