watchman: make win32 boost dep static
Summary:
the same reasons that thrift and folly need to force boost
to be static also apply to watchman: something causes boost to emit
a mixture of static and shared boost library deps when building in
shared library mode.  Without explicitly selecting static mode,
the static deps that get emitted don't have a libdir and thus
cause a linker failure.

Reviewed By: simpkins, strager

Differential Revision: D14726770

fbshipit-source-id: 3b9202bec55386e9c5c8d7dddd4b25e2f2040d07