Add project metadata for local saved state
Summary:
Add a project metadata field to all saved state implementations that
can be used to differentiate multiple saved states for a given project at a
given source control revision. For example, a tool using saved states might
want multiple saved states for a particular revision, one for each supported
version of the tool.

Update the local saved state interface to use the project metadata, if
specified, by appending it, separated by an underscore, to the end of the hash
in the filename.

Reviewed By: wez

Differential Revision: D14789784

fbshipit-source-id: d92d730e4362ff0d8eb1791b840c632ba14be539