Add test for local storage path
Summary:
Add a test to verify that the local storage path for saved state is
being set up properly.

This setup is relatively trivial right now, but it will get more complex as we
add project metadata along with the base project, so it will be handy to have a
test ready for it.

Reviewed By: wez

Differential Revision: D14789137

fbshipit-source-id: e73aee202238c91bf1a3ed92c452809f1df4489f