watchman: fix an issue where a delete wouldn't propagate to its parent
Summary:
This is a slightly thorny thing because the pants build system
(is|was) sensitive to spurious directory changes (see https://github.com/facebook/watchman/issues/305).

sbalabanov noticed that deleting a file didn't also mark its containing
directory as changed.  Digging in, the code that we used to propagate
this was only triggered in the code path where a file was changed,
but not in the code path where a file was deleted.

This diff adds that logic in to the delete path, however, this is
made a little more tricky because simply doing that resulted in
a flakey test.  What is happening is that the filesystem mtime
granularity is not high enough to guarantee that we can observe
the parent directly changing via stat.  The resolution for this
is to tag the request with VIA_NOTIFY so that force delivery of
the parent directory change in the unlink case.

Reviewed By: chadaustin, strager

Differential Revision: D15247692

fbshipit-source-id: 595808fc3291785ec48ae2fc4ef048027e6b02cb