watchman: prevent watching if dir contains an invalid .watchmanconfig
Summary:
We've been bitten by this a couple of times, where someone
edits a config and leaves a trailing comma at the end of an array.
That is invalid json syntax but is hard to detect by eye, and not
every repo is configured to lint that file as json.

This diff throws an exception during watch establishment so that
the issue is more visible.

It is worth noting though that watchman doesn't automatically
reload this config file, so it is still possible for the change
to go unnoticed on the local machine.

Reviewed By: strager

Differential Revision: D15386435

fbshipit-source-id: 9bf6a0205ece24e2fa6dc63e2d1a01116859dcc9