watchman: refuse to run the service at higher than normal nice levels
Summary:
We've had a rash of users run into issues where some automation
ends up spawning source control related services with a lower priority
(aka higher nice level).  The result is massively degraded performance.

To combat this, this diff causes the watchman server to exit with a
reasonable informative error message.  In practice, this error should
only be seen by automation.

When the user later interacts with the service it will spawn as intended.

Reviewed By: simpkins

Differential Revision: D15348126

fbshipit-source-id: 2d08707181147d4f71ac73e6617062b22eaa8624