watchman: really fix notifying parent dirs on macos
Summary:
D15247692 fixed this on Linux, but due to some issues with
our CI on macos, this didn't show as failing on macos until I was
looking at something else.

This diff resolves the issue for mac, which is due to the
gift-that-keeps-giving: case insensitive filesystems.  The propagation check
wasn't triggering on mac because we had already decided to schedule a test to
validate whether the case of the file changed.

The solution is to perform the new propagation check, and if that didn't
schedule an unconditional check of the parent, then perform the speculative
case check.

Reviewed By: strager

Differential Revision: D15346474

fbshipit-source-id: 2f8b6d1389b42e69e9e8e89e6212a8548bb2dcd6