watchman: thrift now requires c++17
Summary:
Our builds were failing on windows because thrift was pulling in some parts of std::optional and that caused MSVC to error out with an error saying that we needed to be enable C++17.   This diff does that.

There's a risk that this will break the travis and appveyor builds, but we'll fix that up once this makes it out to github.

Reviewed By: chadaustin

Differential Revision: D15420267

fbshipit-source-id: 10e44e5c1094270ed2047aaf7a257a6552bdf01a