Fix empty 'path' or 'suffix' generators
Summary:
If a query is made with a `path` or `suffix` generator, and the generator is given an empty array, Watchman acts as if the generator was not specified at all. This behavior is surprising, and it disagrees with the documentation:

> The `all` generator produces a list of all file nodes.  It is the default generator and is used in the case where no other generators were explicitly specified.

(The `glob` generator behaves correctly; querying with `"glob": []` returns no results.)

Fix the `path` and `suffix` generators to produce no results when given an empty array.

Reviewed By: wez

Differential Revision: D15507833

fbshipit-source-id: 8da321a98370bf5ba79b8fbf63a5298d01c02d0c