Kill legacySubscribe method for establishing streams
Summary:
watchman/eden have been using the new Thrift streaming for nearly a year now and are the last users of the old
Thrift streaming. This diff kills the fallback `legacySubscribe` method, which complete's migration to the new Thrift
streaming and unblocks the complete removal of wangle-based streaming from Thrift and fbcode.

Reviewed By: wez

Differential Revision: D15878892

fbshipit-source-id: ec0d270dba79e56c7e41afbf36669a08e5a15518