Eden Mount is_fresh_instance Fix
Summary:
Needed for Eden in the future to return the fact the journal has been truncated.

Also affects code path where EdenServiceHandler throws an ERANGE error, though Watchman gets the mountGeneration earlier to see if this is a fresh instance, if it happened to change after that check then Watchman would ignore Eden's fresh instance.

Reviewed By: wez

Differential Revision: D16036826

fbshipit-source-id: 03a31edb458bb184e80925adc8b9d3ee1d42f2ba