Handle Eden EDOM journal truncation error responses
Summary:
Teach watchman how to process journal truncation error responses generated by Eden.   In order to manage the size of the journal, Eden may decide to truncate older entries.  In those situations Eden will generate EDOM errno values.

In the context of a subscription, take care to avoid terminating the watchman subscription when EDOM is received, as that is a jarring user experience in eg: a buck build, or in Nuclide.

Reviewed By: wez

Differential Revision: D16075256

fbshipit-source-id: 4333674c91946b21c3fef364744387e7eecd615f