Throw a legible exception when watchman get-sockname returns an empty string
Summary: If WatchmanClient is not passed a Watchman socket, it calls `watchman get-sockname` to get a default socket tied to the current unix user. But sometimes `watchman get-sockname` will return an empty string as stdout and an error message to stderr. In these cases, WatchmanClient will attempt to parse the empty string as bser, which will then throw an exception with the message `"underflow"`. I think we can throw a more informative exception here.

Reviewed By: strager

Differential Revision: D16194784

fbshipit-source-id: 3b1c9992d218f3291b4f00f33907548d332b57c8