Subscriptions keep results if is_fresh_instance
Summary: The assumption that a result does nothing because it is empty and doesn't change merge bases does not hold. If "empty_on_fresh_instance" is true then all fresh instances will be empty and then considered useless even though we need to signal to the process watching that they need to handle the fresh instance result in an appropriate manner.

Reviewed By: chadaustin

Differential Revision: D16219267

fbshipit-source-id: 9a0da881c8079bc6d5d23fe4fb9c563a95c34c1b